revamped UI and Cursor systems to handle placing objects on ground
made the UI Decoration Placement script switch which object it's placing based on the button clicks, and then have the MoveSys scripts call the actual decoration placement function. Gonna be fiddling with the UI a bit here, since that's where the decoration spawning script is for some reason.

Feel free to overwrite any of this except the scripts and the UI prefab when pushing changes, btw. I didnt really touch anything bar value adjustment to confirm it works outside of those