overhaul on the backend, plus tweaks
changed various file names to be more clear

modified camera moving script to have variables modifiable in editor, including what button to press to trigger it.

modified UI to make the menu bauble clearly for the menu

modified UI to lack the camera on/off button and the camera switch. one was now useless and the other was broken and not what was asked for.